navigation edit now respects the admin available locales
