+++ upgrade-notes/backend_20250305_064851.md
@@ -1,0 +1,7 @@
+#### make navigation independent on czech domain presence ([#3837](https://github.com/shopsys/shopsys/pull/3837))
+
+- [features moved](#movement-of-features-from-project-base-to-packages) from project-base to the framework packages:
+    - `CategoryFacade::getFullPathsIndexedByIds()` method
+    - `CategoryRepository::getFullPathsIndexedByIds()` method
+    - `CategoryRepository::getPreOrderTreeTraversalForAllCategoriesQueryBuilder()` method
+- see #project-base-diff to update your project
